menambahkan halaman hapus tamu dan fitur hapus data tamu

+++ buku-tamu.php
@@ -79,10 +79,10 @@ if (isset($_POST['simpan'])) {
         <td><?= $tamu['bertemu'] ?></td>
         <td><?= $tamu['kepentingan'] ?></td>
 <td>
-    <a class="btn btn-success" href="edit-tamu.php?id=<?= $tamu['id_tamu'] ?>">Ubah</a>
-    <a class="btn btn-danger" href="hapus-tamu.php?id=<?= $tamu['id_tamu'] ?>" 
-       onclick="return confirm('Yakin ingin hapus?')">Hapus</a>
+  <a class="btn btn-success" href="edit-tamu.php?id=<?= $tamu['id_tamu'] ?>">Ubah</a>
+  <a onclick="return confirm('Apakah anda yakin ingin menghapus data ini?')" class="btn btn-danger" href="hapus-tamu.php?id=<?= $tamu['id_tamu'] ?>">Hapus</a>
 </td>
+
 
     </tr>
     <?php endforeach; ?>
